+++ src/game/base.h
@@ -24,4 +24,4 @@ void base_update(Base *base);
 bool base_hit_test(Base *base, Vec2 hit);
 
 // TODO: handle color conversion in shader
-void base_take_damage(Base *base, Faction from, unsigned amount);
+void base_take_damage(Base *base, Faction from, size_t amount);
